requirements.txt
aiortc
fastapi
uvicorn
websockets
coqui-tts
loguru
numpy
scipy
whisperx
python-multipart
opuslib
mutagen
torch
torchaudio
transformers
accelerate
aiohttp
soundfile
sounddevice
